requirements.txt
streamlit
openai
pyttsx3
SpeechRecognition
pyaudio
python-dotenv
langdetect
